+++ lib/mangoapps.rb
@@ -9,6 +9,10 @@ require_relative "mangoapps/client"
 require_relative "mangoapps/modules/learn"
 require_relative "mangoapps/modules/users"
 require_relative "mangoapps/modules/recognitions"
+require_relative "mangoapps/modules/notifications"
+require_relative "mangoapps/modules/feeds"
+require_relative "mangoapps/modules/posts"
+require_relative "mangoapps/modules/libraries"
 
 module MangoApps; end
 
@@ -16,3 +20,7 @@ module MangoApps; end
 MangoApps::Client.include(MangoApps::Client::Learn)
 MangoApps::Client.include(MangoApps::Client::Users)
 MangoApps::Client.include(MangoApps::Client::Recognitions)
+MangoApps::Client.include(MangoApps::Client::Notifications)
+MangoApps::Client.include(MangoApps::Client::Feeds)
+MangoApps::Client.include(MangoApps::Client::Posts)
+MangoApps::Client.include(MangoApps::Client::Libraries)