+++ lib/mangoapps/modules/notifications.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require_relative "notifications/my_priority_items"
+
+module MangoApps
+  class Client
+    module Notifications
+      # Include all notifications sub-modules
+      include MangoApps::Client::Notifications::MyPriorityItems
+    end
+  end
+end